move 'IconSelect' component into 'IconSelect' file
